+++ src/main/java/lotto/Application.java
@@ -1,9 +1,11 @@
 package lotto;
 
+import lotto.controller.LottoController;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        Controller controller = new Controller();
+        LottoController controller = new LottoController();
         controller.run();
     }
 }
